+++ app/src/SettingsTabs/BackgroundColor.css
@@ -2,4 +2,5 @@
   color: red;
   padding-left: 10px;
   font-weight: bold;
+  display: inline;
 }
