+++ app/src/SettingsTabs/BackgroundColor.tsx
@@ -1,42 +1,43 @@
 import { observer } from "mobx-react-lite";
-import React from "react";
+import React, {useContext} from "react";
 import { FormLabel, Typography } from "@mui/material";
 import TSCColorPicker from "../components/TSCColorPicker";
-import { setRGB } from "../state/actions";
 import "./BackgroundColor.css";
-import { convertHexToRGB } from "../state/actions/util-actions";
+import { convertHexToRGB } from "../util/util";
 import { ColumnInfo } from "@tsconline/shared";
+import { context } from "../state";
 
 interface ChangeBGColorProps {
   column: ColumnInfo;
 }
 export const ChangeBackgroundColor: React.FC<ChangeBGColorProps> = observer(({ column }) => {
+  const { state, actions } = useContext(context);
   const handleColorChange = (color: string) => {
-    setRGB(column, convertHexToRGB(color, false));
+    actions.setRGB(column, convertHexToRGB(color, false));
   };
 
   if (column.children.length != 0) {
     return (
-      <div>
-        <FormLabel id="color-label" className="bg-label">
+      <>
+        <FormLabel>
           Background Color:
-          <Typography className="not-avail" style={{ display: "inline" }}>
+          <Typography className="not-avail">
             Not Available
           </Typography>
         </FormLabel>
-      </div>
+      </>
     );
   } else {
     return (
       <div>
-        <FormLabel id="color-label" className="bg-label">
+        <FormLabel>
           Background Color:&nbsp;
         </FormLabel>
-        <TSCColorPicker
-          key={column.name}
-          color={`rgb(${column.rgb.r}, ${column.rgb.g}, ${column.rgb.b})`}
-          onColorChange={handleColorChange}
-        />
+          <TSCColorPicker
+              key={column.name}
+              color={`rgb(${column.rgb.r}, ${column.rgb.g}, ${column.rgb.b})`}
+              onColorChange={handleColorChange}
+          />
       </div>
     );
   }
